refactor: WebConfig만 빈으로 등록하도록 수정
