+++ src/main/java/koreatech/in/config/WebConfig.java
@@ -1,4 +1,4 @@
-package koreatech.in.converter;
+package koreatech.in.config;
 
 import java.util.Collections;
 
@@ -7,6 +7,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.support.ConversionServiceFactoryBean;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
+import koreatech.in.converter.DomainEnumConvertor;
 
 @Configuration
 @EnableWebMvc
